more robustly load config file
Apparently when I launch emacs from kde's launcher, there is no buffer
file name. This means that ~/readme.el is what it attempts to load,
but the right filepath is ~/.emacs.d/readme.el

I'd like to test this since it's a very foundational problem, but I
can't think of a good way to test it on CI atm short of installing kde
and setting up xvrb, lol.