make sure private work/personal files are loaded
depends on env var

This used to just be a link to personal.nix or work.nix respectively,
but there's no reason not to keep a very very large amount of both of
those categories tracked in this git repo.

The remainder I'll put in these private files, but I found without
thinking of them as "for private only" they grew quite large and
undermined the reproducability of the system as a whole.