improve performance by stuffing everything into an async block
In theory this won't block and the callback will set the variable
correctly. Then with basically every command updating since we hook
post-command-hook it should update quickly?

Sadly I had to stuff everything into the lambda that gets sent to the
other external emacs to execute... but... if it works ¯\_(ツ)_/¯