don't use tab for yasnippet expansion
I'm not sure about this one. I've had problems with tab completion in
the past anyway. The alternative to this would be to not use
company-tng (not sure what this would be like) and then test to see if
I yasnippets just work correctly.

I believe I had other problems with them too though.